Updated PHP frameworks for Windows
Fixed PHP installation on Windows: a new version of PHP was released and the previous version was moved to http://windows.php.net/downloads/releases/archives/ I switched the url to archives (instead of using the new version) because otherwise the installer script will be broken every time a new PHP version is released.

+++ cake/setup.py
@@ -1,6 +1,7 @@
 
 import subprocess
 import sys
+import os
 import setup_util
 from os.path import expanduser
 
@@ -13,6 +14,11 @@ def start(args):
   setup_util.replace_text("cake/deploy/nginx.conf", "root .*\/FrameworkBenchmarks", "root " + home + "/FrameworkBenchmarks")
 
   try:
+    if os.name == 'nt':
+      setup_util.replace_text("cake/app/Config/core.php", "'Apc'", "'Wincache'")
+      subprocess.check_call('icacls "C:\\FrameworkBenchmarks\\cake" /grant "IIS_IUSRS:(OI)(CI)F"', shell=True)
+      subprocess.check_call('appcmd add site /name:PHP /bindings:http/*:8080: /physicalPath:"C:\\FrameworkBenchmarks\\cake\\app\\webroot"', shell=True)
+      return 0
     #subprocess.check_call("sudo cp cake/deploy/cake /etc/apache2/sites-available/", shell=True)
     #subprocess.check_call("sudo a2ensite cake", shell=True)
     subprocess.check_call("sudo chown -R www-data:www-data cake", shell=True)
@@ -24,6 +30,9 @@ def start(args):
     return 1
 def stop():
   try:
+    if os.name == 'nt':
+      subprocess.call('appcmd delete site PHP', shell=True)
+      return 0
     subprocess.call("sudo /usr/local/nginx/sbin/nginx -s stop", shell=True)
     subprocess.call("sudo kill -QUIT $( cat cake/deploy/php-fpm.pid )", shell=True)
     #subprocess.check_call("sudo a2dissite cake", shell=True)
